# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtefactKind": {
        "description": "What a stored file is. The bytes never enter the database.",
        "enum": [
          "report.html",
          "report.pdf",
          "run.json",
          "ranking.csv"
        ],
        "title": "ArtefactKind",
        "type": "string"
      },
      "ArtefactOut": {
        "additionalProperties": false,
        "description": "A file belonging to a run, as a client may see it.\n\n`uri` is not here. It is a path on the server's disk and a client has no use for\none; `href` is the URL that serves the bytes, and `sha256` is what they should hash\nto when they arrive.",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "href": {
            "title": "Href",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ArtefactKind"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "run_id",
          "kind",
          "size_bytes",
          "sha256",
          "created_at",
          "href"
        ],
        "title": "ArtefactOut",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "The response body of every refusal this API makes.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Refusal"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "Refusal": {
        "additionalProperties": false,
        "description": "Why the request was refused, in the terms the caller can act on.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "errors": {
            "description": "Every fault found, when a request had more than one.",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The single thing at fault, when there is one.",
            "title": "Field"
          },
          "message": {
            "description": "Written for a person. Names the column, the id or the setting.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "Refusal",
        "type": "object"
      },
...
      "Run": {
        "additionalProperties": false,
        "description": "A finished assessment: the whole payload, plus the few things worth indexing.\n\nThe payload is stored entire and is the only source of truth. Every other column\nhere is a copy of something inside it, lifted out so a list of runs can be drawn\nwithout opening any of them \u2014 and each one is written from the payload on insert,\nnever supplied separately, so they cannot disagree with it.",
        "properties": {
          "corridor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corridor Id"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "rung": {
            "title": "Rung",
            "type": "string"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "project_id",
          "engine_version",
          "fingerprint",
          "mode",
          "rung",
          "payload"
        ],
        "title": "Run",
        "type": "object"
      },
      "RunSummary": {
        "additionalProperties": false,
        "description": "A run without its payload. What a listing is made of.\n\nEvery field is one the store lifted out of the payload on insert, so a summary\ncannot describe a different run than the one it points at.",
        "properties": {
          "corridor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corridor Id"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "rung": {
            "title": "Rung",
            "type": "string"
          },
          "schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schema Version"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "project_id",
          "job_id",
          "corridor_id",
          "schema_version",
          "engine_version",
          "fingerprint",
          "mode",
          "rung",
          "created_at"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nCorridor road-risk assessment from open data, with provenance.\n\n**Two things this deployment does not yet do, stated here rather than discovered.**\n\n* **Nothing executes jobs.** `POST /jobs` stores a job and returns `202`. It stays\n  `queued` until step 5.1d attaches a runner. `GET /health` reports `runner: null`.\n* **`X-Tenant-Id` is not authentication.** It scopes every read to one tenant, which is\n  what keeps two clients' runs apart, but nothing verifies the claim. Step 5.4a\n  replaces it with real identities and row-level policies in the database.\n\n**A refusal is a result, not an error.** A panel that breaks the input contract is a\n`422` naming the column, and no job is created. An assessment that descended to Mode B,\ndropped a term or declined to score an unsourced weight is a **`200`** \u2014 those are\nfindings the run carries, not failures. Infrastructure breaking is a job whose status is\n`failed`, with a cause.\n\nPayload shape version `1.0`. Factors, tiers and licences come from\n`factors.yaml`; `GET /registry` serves it with the hash of the file it was read from.\n",
    "summary": "Corridor risk from open data, with every number traceable.",
    "title": "Road Risk Panel",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs": {
      "get": {
        "description": "Summaries, not payloads.\n\nA run is around 300 kB of JSON. Fifty of them is fifteen megabytes, which is not a\nlisting \u2014 it is a download nobody asked for. Every field in a summary was lifted out\nof the payload by the store on insert, so a summary cannot describe a different run\nthan the one it points at.",
        "operationId": "list_runs_runs_get",
        "parameters": [
          {
            "description": "Narrow to one project. Omitted lists every project.",
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Narrow to one project. Omitted lists every project.",
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response List Runs Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request or the panel was refused."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          }
        },
        "summary": "Finished runs, newest first, without their payloads",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}": {
      "get": {
        "description": "Return the run and its payload.\n\n**200 whatever the assessment concluded.** A run that descended to Mode B carries\nits descent receipt, the terms it dropped and the weights it would not score, and\nall of that is inside `payload`. Turning any of it into a status code would move\nthe honesty layer into an error handler, which is where nobody reads it.\n\nThe payload is not re-validated on the way out. `store_run` validated it on the way\nin and there is nowhere for it to have changed since \u2014 validating again would be a\nsecond answer to a question already answered, and the two could disagree.",
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request or the panel was refused."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          }
        },
        "summary": "One run, whole \u2014 including everything it refused to conclude",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/artefacts": {
      "get": {
        "description": "List them with a URL, never with the path they sit at on this server.",
        "operationId": "list_artefacts_runs__run_id__artefacts_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtefactOut"
                  },
                  "title": "Response List Artefacts Runs  Run Id  Artefacts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request or the panel was refused."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          }
        },
        "summary": "The files this run produced",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/artefacts/{kind}": {
      "get": {
        "description": "Serve the file, having proved it is one this server is allowed to open.\n\nThe `ETag` is the sha256 the store recorded, so a client can verify what arrived\nagainst what was registered without this route re-hashing a third of a megabyte on\nevery request. The size *is* checked, because `stat` is free and a file whose length\nno longer matches the record is not the artefact the record describes.",
        "operationId": "download_artefact_runs__run_id__artefacts__kind__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ArtefactKind"
            }
          },
          {
            "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Which tenant's rows this request may see. Required on every route that touches a row. **This is not authentication** \u2014 step 5.4a replaces it with a real identity and row-level policies in the database. An unknown tenant is not an error: it is a tenant with no rows, which is what tenancy means.",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The bytes. `ETag` is the sha256 recorded at import."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Refused before anything was read."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No tenant. See X-Tenant-Id."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "No such row, for this tenant."
          },
          "409": {
            "description": "Recorded, but this server will not or cannot serve it \u2014 no artefact root configured, outside it, missing, or changed since."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "The request or the panel was refused."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Logged, with a reference. Never a trace."
          },
          "501": {
            "description": "Stored somewhere this build cannot fetch from."
          }
        },
        "summary": "Download one artefact",
        "tags": [
          "runs"
        ]
      }
    }
  }
}